Fix PKHeX.Avalonia compile errors on Linux
- CurrentLevel is a byte in Core; cast Level before assigning
- Qualify Avalonia.Application/Controls with global:: to avoid clash
  with the PKHeX.Avalonia namespace
- Drop WithInterFont() (requires the separate Avalonia.Fonts.Inter
  package); the default font is sufficient

+++ PKHeX.Avalonia/ViewModels/PokemonEditorViewModel.cs
@@ -160,7 +160,7 @@ public partial class PokemonEditorViewModel : ViewModelBase
 
         // Write all editable fields back to the PKM.
         _pk.Nickname = Nickname;
-        _pk.CurrentLevel = Level;
+        _pk.CurrentLevel = (byte)Level;
         _pk.OriginalTrainerName = OtName;
         _pk.IsNicknamed = IsNicknamed;
